feat: clean the evaluation api
